feat: Helmfile renders *.yaml.gotmpl in a K8s manifests/kustomization directory
Related to https://github.com/roboll/helmfile/issues/494

This feature is mostly a built-in alternative to incubator/raw chart without external dependency and has
access to helmfile's own template functions and template data.

The expected use-case of this feature is to add arbitrary K8s resources to your deployment.

Unlike the original issue raised in #494 this doesn't enable you to add arbitary resources to a release. That's another story. But this would be a good foundation for that, too.